+++ _assets/globals.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    adminBarPostRequest(target: HTMLElement, request: string, body: string)
+  }
+}
+
+export {}